tools/telemetry/litebrain/requirements.txt
fastapi>=0.98.0,<0.120.0
uvicorn>=0.22.0,<1.0.0
opentelemetry-api>=1.10.0
opentelemetry-sdk>=1.10.0
opentelemetry-instrumentation==0.35b0
opentelemetry-instrumentation-fastapi==0.35b0
opentelemetry-instrumentation-requests==0.35b0
opentelemetry-exporter-otlp-proto-grpc>=1.10.0
requests>=2.31.0
